cmd: move runinhibit to snaplock
